requirements.txt
click==8.1.3
PyYAML==6.0
telebot==0.0.5
vk-api==11.9.9
python-dotenv==1.0.0
